Fix: edge case for transplanting multiline json string (#102)
* Fix: edge case for transplanting multiline json string

When trying to copy a multiline json string from a yaml to a string in another yaml, there are issues with the JSON being parsed during the initial get which is undesireable. if this edge case applies to you, you can use this new type jsonString to let us know it is a json string that you want to stay a json string at its destination. we will force the quotes back to the proper escape level.

* Build: npm audit-ci breaking the build, going back to npm audit