fix: dont null whole objects if a resource stops creating it (#194)
* fix: dont null whole objects if a resource stops creating it

this bug was found when a user had a MTP create a resource with an annotation section defined. They then updated their mtp and no longer defined the annotation section. This caused the reconcileFields function to null the entire annotations section. We then tell objectPath to null out the old kapitan/last-applied annotation, and it promply fails because it can set a value to a null field.

Not only will the commit fix that user issue, it will also prevent us destroying other items in a resource that a user may stop defining but may not what the whole thing deleted (like some default set by kube or a third party)

* fix: allow errors to propogate to resource's status