fix: dont let annotations be null after reconcile
if a user defines annotations, and then in the future removes all annotations, reconcileFields will null out all of the annotations object. Since we need annotations to exist, and object path is incapable of writing to a path with null in part of the path, we need to ensure annotations is not null before writing our razee last applied out.